hm2_eth: use read deadline for queued reads
An up/down error timer is added so that one failure doesn't signal
an error to the upper layers.  Instead, enough errors over time have
to happen so that the counter reaches 10.  Ultimately, this will need
to be a tunable parameter, and output pins will be needed to monitor
packet loss.

Packet loss can be simulated by iptables.  The following script,
when executed as root, causes .001 of all packets to be lost.
Executing it again with any argument disables packet loss.

Losing .1% of all packets means that .2% of all reads will fail,
since they consist of two packets (request and response).

    #!/bin/sh
    set -xe
    iptables -n -L loss-in 2>/dev/null || iptables -N loss-in
    iptables -n -L loss-out 2>/dev/null || iptables -N loss-out
    iptables -C OUTPUT -j loss-out 2>/dev/null || iptables -I OUTPUT 1 -j loss-out
    iptables -C INPUT -j loss-in 2>/dev/null || iptables -I INPUT 1 -j loss-in
    if [ $# -eq 0 ]; then
    iptables -A loss-out -i eth1  -m statistic --mode random --probability .001 -j DROP
    iptables -A loss-in -i eth1  -m statistic --mode random --probability .001 -j DROP
    else
    iptables -F loss-in
    iptables -F loss-out
    fi

With this level of packet loss, hostmot2 frequently fails to start.
But if the packet loss is added only after startup, the behavior of
other parts of the system in the face of packet loss can begin to be
inspected.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>